+++ src/components/Dialog/Modal.tsx
@@ -23,11 +23,14 @@ export interface ModalProps extends OverlayProps, IOverlayState {
   preventBackdrop?: boolean;
     /** Optional props for the wrapper div element inside the modal overlay. */
   wrapperDivProps?: React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement>
+    /** Make the modal focusable, e.g. when clicking somewhere on it. This is needed, e.g. when capturing key (down, up) events that
+     * should bubble to the modal's parent elements. */
+    modalFocusable?: boolean
 }
 
 /**
  * Displays contents on top of other elements, used to create dialogs.
- * For most situations the usage of `SimpleDialog` and `AlertDialog` should be sufficent.
+ * For most situations the usage of `SimpleDialog` and `AlertDialog` should be sufficient.
  * Otherwise this element can be used to create own modal elements and edge cases for modal dialogs.
  * Then it is recommended to use the `Card` element inside.
  */
@@ -40,8 +43,19 @@ export const Modal = ({
     canEscapeKeyClose=false,
     preventBackdrop=false,
     wrapperDivProps,
+    modalFocusable = true,
     ...otherProps
 }: ModalProps) => {
+
+    const backdropProps: React.HTMLProps<HTMLDivElement> | undefined = !canOutsideClickClose && canEscapeKeyClose ? {
+        ...otherProps.backdropProps,
+        // Escape key won't work anymore otherwise after clicking on the backdrop
+        tabIndex: 0
+    } : otherProps.backdropProps
+
+    const focusableProps = modalFocusable ? {
+        tabIndex: 0
+    } : undefined
 
     const alteredChildren = React.Children.map(children, (child) => {
         if ((child as React.ReactElement).type && (child  as React.ReactElement).type === Card) {
@@ -60,6 +74,7 @@ export const Modal = ({
     return (
         <BlueprintOverlay
             {...otherProps}
+            backdropProps={backdropProps}
             className={overlayClassName}
             backdropClassName={`${eccgui}-dialog__backdrop`}
             canOutsideClickClose={canOutsideClickClose}
@@ -71,6 +86,7 @@ export const Modal = ({
                 className={BlueprintClassNames.DIALOG_CONTAINER}
                 // this is a workaround because data attribute on SimpleDialog is not correctly routed to the overlay by blueprint js
                 data-test-id={(otherProps as any)["data-test-id"] ?? "simpleDialogWidget"}
+                {...focusableProps}
             >
                 <section
                     className={